rename module brent to scalar
