+++ src/polykin/math/optimization/scalar.py
